MeiTuKanKan
This is an Android project I did with two other students in the second semester of my sophomore year after learning the basic course of Android. We wanted to make it possible for users to see the images that we have on the server on the app, and also to bookmark and download the images. We did this project in a very rough way, and we finished it in only five days.